新增微信监控

+++ video/jianying_fengmian_gap.py
@@ -22,6 +22,11 @@ if not GuiOpt.find_icon(add_url("jianying_bianjicaidan.png")):
     GuiOpt.click_icon(add_url("jianying_icon.png"))
 # GuiOpt.click_icon(add_url("jianying_icon2.png")
 GuiOpt.wait_appear(add_url("jianying_fengmian.png"))
+# 如已有封面则删除
+if not GuiOpt.find_icon(add_url("jianying_fengmian.png")):
+    GuiOpt.move_to(160, 808)
+    GuiOpt.click_pos(160, 808)
+    GuiOpt.enter()
 GuiOpt.click_icon(add_url("jianying_fengmian.png"))
 GuiOpt.wait_appear(add_url("jianying_bendi.png"), 10)
 GuiOpt.click_icon(add_url("jianying_bendi.png"))
@@ -65,23 +70,30 @@ GuiOpt.paste()
 GuiOpt.click_icon(add_url("jianying_wanchengshezhi.png"))
 
 # 导入图片
-if CommonApi.get_os_type() == "win":
-    srcpath = GlobalVar.get("imgsavedir")
-    destpath = "C:\\Users\\Administrator\\Desktop\\temp_del"
-    if os.path.exists(destpath):
-        print("存在同名目录，正在删除: " + destpath)
-        CommonApi.remove_directory(destpath)
-    else:
-        os.mkdir(destpath)
-    CommonApi.copy_files(srcpath, destpath)
-    GuiOpt.click_icon(add_url("jianying_daoru.png"))
-    GuiOpt.click_icon(add_url("jianying_zhuomian.png"))
-    GuiOpt.double_click_icon(add_url("jianying_temp_del.png"))
-    GuiOpt.select_all()
-    GuiOpt.click_icon(add_url("jianying_dakai2.png"))
+if GuiOpt.find_icon(add_url("jianying_daoru.png")):
+    if CommonApi.get_os_type() == "win":
+        srcpath = GlobalVar.get("imgsavedir")
+        destpath = "C:\\Users\\Administrator\\Desktop\\temp_del"
+        if os.path.exists(destpath):
+            print("存在同名目录，正在删除: " + destpath)
+            CommonApi.remove_directory(destpath)
+        else:
+            os.mkdir(destpath)
+        CommonApi.copy_files(srcpath, destpath)
+        GuiOpt.click_icon(add_url("jianying_daoru.png"))
+        if not GuiOpt.find_icon(add_url("jianying_daoruzuidahuachuangkou.png")):
+            GuiOpt.double_click_icon(add_url("jianying_xuanzemeitiziyuan.png"))
+        GuiOpt.click_icon(add_url("jianying_zhuomian.png"))
+        GuiOpt.double_click_icon(add_url("jianying_temp_del.png"))
+        GuiOpt.select_all()
+        GuiOpt.click_icon(add_url("jianying_dakai2.png"))
+else:
+    print("跳过导入")
 
 # 插入结尾
 GuiOpt.click_icon(add_url("jianying_shipin_jiewei.png"))
 GuiOpt.click_icon(add_url("jianying_wodeyushe.png"))
-GuiOpt.click_icon(add_url("jianying_pianwei.png"))
-GuiOpt.click_icon(add_url("jianying_tianjiayushe.png"))
+# GuiOpt.click_icon(add_url("jianying_pianwei.png")
+GuiOpt.click_icon(add_url("jianying_pianwei2.png"))
+# GuiOpt.click_icon(add_url("jianying_tianjiayushe.png"))
+GuiOpt.click_icon(add_url("jianying_tianjiayushe2.png"))
